Use the new `SCCACHE_PID_FILE` envvar so the sccache daemon doesn't randomly change pids (#709)

+++ features/src/utils/opt/devcontainer/bin/sccache/start.sh
@@ -58,15 +58,21 @@ _start_sccache() {
     else
         # Start the sccache server in the background
         RUST_LOG_STYLE="never"                  \
+        SCCACHE_PID_FILE="${pidfile}"           \
         SCCACHE_ERROR_LOG="${logfile}"          \
         SCCACHE_SERVER_LOG="${log_lvl}"         \
         SCCACHE_SERVER_PORT="${sccache_port}"   \
         sccache --start-server 1>&2 2>/dev/null \
       | tee "$logfile";
-        # Write the pid to the pidfile
-        pgrep sccache | sort -n | head -n1 | tee "${pidfile}" >/dev/null;
+
+        # If the pidfile doesn't exist, write it
+        if ! test -s "${pidfile}"; then
+            pgrep sccache | sort -n | head -n1 | tee "${pidfile}" >/dev/null;
+        fi
+
         # Increase the open file limit so users can do `make -j(ulimit -n)`
         prlimit --nofile=$(ulimit -Hn):$(ulimit -Hn) --pid "$(cat "${pidfile}")";
+
         echo "=== sccache logfile: $logfile ===" >&2;
         echo "=== sccache pidfile: $pidfile ===" >&2;
     fi
